+++ css/added-style.css
@@ -17,7 +17,27 @@
 }
 
 .error-block {
-  width: 50%;
-  height: 50%;
-  border: 5px solid blueviolet;
+  position: absolute;
+  top: 100px;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 70%;
+  height: 50px;
+  background-color: white;
+  border: 4px solid #ffaa99;
+  border-radius: 8px;
+  font-size: 22px;
+  font-weight: 700;
+  z-index: 2;
+}
+
+.map {
+  z-index: 1;
+}
+
+.hidden {
+  display: none;
 }
