--- js/get-similar-adds.js
@@ -1,6 +1,0 @@
-import {createAdds} from './data.js';
-import {SIMILAR_ADD_COUNT} from './preset-const.js';
-
-const similarAdds = Array.from({length: SIMILAR_ADD_COUNT}, createAdds);
-
-export {similarAdds};
