+++ js/validate-form.js
@@ -1,39 +1,9 @@
 import './libs/noUiSlider-init.js';
 import {pristine} from './libs/pristin-init.js';
-import {form, range, capacityField, roomsField, roomsOption, guestsOption, valuesDivs, roomsGuestsErrorMessage,
+
+import {range, capacityField, roomsField, roomsOption, guestsOption, valuesDivs, roomsGuestsErrorMessage,
   typeField, priceField, typesMinPrice, timeFieldset, formInputs} from './const.js';
-import {blockSubmitButton, deleteErrors, setValue, unblockSubmitButton} from './utils.js';
-import {sendData} from './api.js';
-
-// Валидируем поля с помощью Пристин
-export const setUserFormSubmit = (onSuccess, onFail) => {
-
-  form.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-    pristine.validate();
-
-    if(pristine.validate()) {
-
-      blockSubmitButton();
-
-      sendData(
-
-        () => {
-          onSuccess();
-          unblockSubmitButton();
-        },
-
-        () => {
-          onFail();
-          unblockSubmitButton();
-        },
-
-        new FormData(evt.target)
-
-      );
-    }
-  });
-};
+import {deleteErrors, setValue} from './utils.js';
 
 // Валидируем число комнат и гостей
 const validateRooms = () => roomsOption[roomsField.value].includes(capacityField.value);
